Rework and fix transitions in build for v2
