+++ examples/using-page-transitions/gatsby-browser.js
@@ -1,10 +1,5 @@
-/* eslint-disable react/prop-types */
 /* globals window CustomEvent */
-import React, { createElement } from "react"
-import { Transition } from "react-transition-group"
 import createHistory from "history/createBrowserHistory"
-
-import getTransitionStyle from "./src/utils/getTransitionStyle"
 
 const timeout = 250
 const historyExitingEventType = `history::exiting`
@@ -26,68 +21,4 @@ if (typeof document !== 'undefined') {
 
 export let replaceHistory = () => history
 
-class ReplaceComponentRenderer extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = { exiting: false, nextPageResources: {} }
-    this.listenerHandler = this.listenerHandler.bind(this)
-  }
-
-  listenerHandler(event) {
-    const nextPageResources = this.props.loader.getResourcesForPathname(
-      event.detail.pathname,
-      nextPageResources => this.setState({ nextPageResources })
-    ) || {}
-    this.setState({ exiting: true, nextPageResources })
-  }
-
-  componentDidMount() {
-    window.addEventListener(historyExitingEventType, this.listenerHandler)
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener(historyExitingEventType, this.listenerHandler)
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (this.props.location.key !== nextProps.location.key) {
-      this.setState({ exiting: false, nextPageResources: {} })
-    }
-  }
-
-  render() {
-    const transitionProps = {
-      timeout: {
-        enter: 0,
-        exit: timeout,
-      },
-      appear: true,
-      in: !this.state.exiting,
-      key: this.props.location.key,
-    }
-    return (
-      <Transition {...transitionProps}>
-      {
-        (status) => createElement(this.props.pageResources.component, {
-          ...this.props,
-          ...this.props.pageResources.json,
-          transition: {
-            status,
-            timeout,
-            style: getTransitionStyle({ status, timeout }),
-            nextPageResources: this.state.nextPageResources,
-          },
-        })
-      }
-      </Transition>
-    )
-  }
-}
-
-// eslint-disable-next-line react/display-name
-export let replaceComponentRenderer = ({ props, loader }) => {
-  if (props.layout) {
-    return undefined
-  }
-  return createElement(ReplaceComponentRenderer, { ...props, loader })
-}
+export { historyExitingEventType, timeout }
